fix: rename GetIdentityByAddress to GetIdentity
